Enhances map marker functionality
Improves the map component by adding support for multiple markers with customizable icons.

This allows users to add, remove and filter markers on the map, enhancing user interaction and information display.
The component now supports place search and annotation.

The parent place is passed as a property, and the component is initialized based on this prop.